backend/requirements.txt
# BeccaBot MVP Backend - Python 3.14 compatible (no ChromaDB)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17

# Embeddings (OpenAI)
openai>=1.55.0

# Document parsing
pypdf>=5.1.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0

# Web fetching for links
requests>=2.31.0
beautifulsoup4>=4.12.0

# Config
python-dotenv>=1.0.0
